chore: refactoring withdraw modal
