+++ apps/frontend/src/app/5_pages/MarketMakingPage/components/BobWIthdrawModal/hooks/useHandleSubmit.ts
@@ -14,11 +14,11 @@ import { useAccount } from '../../../../../../hooks/useAccount';
 import { useCurrentChain } from '../../../../../../hooks/useChainStore';
 import { translations } from '../../../../../../locales/i18n';
 import { PoolPositionType } from '../../../MarketMakingPage.types';
+import { useGetPool } from '../../../hooks/useGetPool';
 import { AmbientPosition } from '../../AmbientMarketMaking/AmbientMarketMaking.types';
 import { AmbientLiquidityPool } from '../../AmbientMarketMaking/utils/AmbientLiquidityPool';
 import { AmbientLiquidityPoolDictionary } from '../../AmbientMarketMaking/utils/AmbientLiquidityPoolDictionary';
 import { DEFAULT_SLIPPAGE } from '../../BobDepositModal/BobDepositModal.constants';
-import { useGetPoolInfo } from './useGetPoolInfo';
 
 export const useHandleSubmit = (
   withdrawAmount: BigNumber,
@@ -29,7 +29,7 @@ export const useHandleSubmit = (
 ) => {
   const { signer } = useAccount();
   const { croc } = useCrocContext();
-  const { poolTokens } = useGetPoolInfo(pool.base, pool.quote);
+  const { poolTokens } = useGetPool(pool.base, pool.quote);
 
   const chainId = useCurrentChain();
   const { setTransactions, setIsOpen, setTitle } = useTransactionContext();
